refactor: Modularize labels context
