+++ github-issue/src/components/Labels/LabelsContext.js
@@ -1,0 +1,3 @@
+import React from "react";
+
+export const LabelsContext = React.createContext();
